VFB-409: refactored old table file and imports
