+++ src/app/clients/clientsTable/headers.ts
@@ -1,4 +1,4 @@
-import { TableHeaders } from "@/components/Tables/Table";
+import { TableHeaders } from "@/components/Tables/materialTable/tableTypes";
 import { ClientsTableRow } from "./types";
 
 const clientsHeaders: TableHeaders<ClientsTableRow> = [
